+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Migrate regression classes into regression module
 * Add split conformal option for regression and classification
 * Update check method for calibration
+* Fix bug in MapieClassifier with different number of labels in calibration dataset
 
 0.6.4 (2023-04-05)
 ------------------