+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 ##### (##########)
 ------------------
 
+* Add grouped conditional coverage metrics named SSC for regression and classification
+* Add HSIC metric for regression
 * Migrate conformity scores classes into conformity_scores module
 * Migrate regression classes into regression module
 * Add split conformal option for regression and classification